Restructured files in the solution
-added models into model namespace
-added repository into repository namespace
