+++ GraduationTracker/GraduationTracker/GraduationTracker.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using GraduationTracker.Models;
+using GraduationTracker.Repositories;
 
 namespace GraduationTracker
 {
