+++ GraduationTracker/GraduationTracker/GraduationTracker.csproj
@@ -40,14 +40,18 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Course.cs" />
-    <Compile Include="Diploma.cs" />
+    <Compile Include="Models\Course.cs" />
+    <Compile Include="Models\Diploma.cs" />
     <Compile Include="GraduationTracker.cs" />
-    <Compile Include="STANDING.cs" />
+    <Compile Include="Models\STANDING.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Repository.cs" />
-    <Compile Include="Requirement.cs" />
-    <Compile Include="Student.cs" />
+    <Compile Include="Repositories\Repository.cs" />
+    <Compile Include="Models\Requirement.cs" />
+    <Compile Include="Models\Student.cs" />
+    <Compile Include="Services\DiplomaService.cs" />
+    <Compile Include="Services\IDiplomaRequirementService.cs" />
+    <Compile Include="Services\IRequirementService.cs" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
