+++ GraduationTracker/GraduationTracker/Models/Course.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GraduationTracker
+﻿namespace GraduationTracker.Models
 {
     public class Course
     {
@@ -12,5 +6,5 @@ namespace GraduationTracker
         public string Name { get; set; }
         public int Mark { get; set; }
         public int Credits { get; }
-     }
+    }
 }
