+++ GraduationTracker/GraduationTracker/Models/Diploma.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GraduationTracker
+﻿namespace GraduationTracker.Models
 {
     public class Diploma
     {
